Add preset chips in reminder sheet. Refactor snooze activity to use common function.
